color and design changed, ready for assesment

+++ js/index.js
@@ -27,7 +27,7 @@ const seatContainer = document.getElementById("seatContainer");
         (seatContainer.innerHTML += `
          <div 
             id="${seat.id}" 
-            class="${seat.status === 1 ? "bg-success" : "bg-danger"}  ${
+            class="${seat.status === 1 ? "availableColor" : "occupiedColor"}  ${
             seat.booked && seat.status ? "border-4 border border-warning" : ""
         } p-2 me-2 mb-2" style="width:calc(100% * (1/3) - 10px - 1px) "
             onclick="addSeat(${seat.id})"
@@ -77,30 +77,35 @@ function bookSeat() {
     location.reload();
 }
 function cancelSeat() {
-    if (!localStorageSeats) {
-        alert("No seat is selected");
-        return;
-    }
-    const occupied = localStorageSeats.filter((item) => item.status === 0);
-    const booked = localStorageSeats.filter((item) => item.booked);
-    if (booked.length === 0) {
-        alert("No seat is selected");
-        return;
-    }
-    if (occupied.length === 0) {
-        alert("No seat is selected");
-        return;
-    }
+    const confirm = window.confirm(
+        "Clicking this, All tickets will be restored to available."
+    );
+    if (confirm) {
+        if (!localStorageSeats) {
+            alert("No seat is selected");
+            return;
+        }
+        const occupied = localStorageSeats.filter((item) => item.status === 0);
+        const booked = localStorageSeats.filter((item) => item.booked);
+        if (booked.length === 0) {
+            alert("No seat is selected");
+            return;
+        }
+        if (occupied.length === 0) {
+            alert("No seat is booked");
+            return;
+        }
 
-    const selectedSeats = localStorageSeats.forEach((item) => {
-        item.booked = false;
-        item.status = 1;
-        item.user = null;
-    });
+        const selectedSeats = localStorageSeats.forEach((item) => {
+            item.booked = false;
+            item.status = 1;
+            item.user = null;
+        });
 
-    const newSeat = [...localStorageSeats];
-    localStorage.setItem("seats", JSON.stringify(newSeat));
-    location.reload();
+        const newSeat = [...localStorageSeats];
+        localStorage.setItem("seats", JSON.stringify(newSeat));
+        location.reload();
+    }
 }
 
 //logout
